Rename Test project & update some libraries
Rename Test project & update some libraries
