+++ .vscode/tasks.json
@@ -7,7 +7,7 @@
             "type": "process",
             "args": [
                 "build",
-                "${workspaceFolder}/test/ReplaceDateTimeNowTest/ReplaceDateTimeNowTest.csproj",
+                "${workspaceFolder}/test_sample/DemoWebApp/DemoWebApp.csproj",
                 "/property:GenerateFullPaths=true",
                 "/consoleloggerparameters:NoSummary"
             ],
@@ -19,7 +19,7 @@
             "type": "process",
             "args": [
                 "publish",
-                "${workspaceFolder}/test/ReplaceDateTimeNowTest/ReplaceDateTimeNowTest.csproj",
+                "${workspaceFolder}/test_sample/DemoWebApp/DemoWebApp.csproj",
                 "/property:GenerateFullPaths=true",
                 "/consoleloggerparameters:NoSummary"
             ],
@@ -32,7 +32,7 @@
             "args": [
                 "watch",
                 "run",
-                "${workspaceFolder}/test/ReplaceDateTimeNowTest/ReplaceDateTimeNowTest.csproj",
+                "${workspaceFolder}/test_sample/DemoWebApp/DemoWebApp.csproj",
                 "/property:GenerateFullPaths=true",
                 "/consoleloggerparameters:NoSummary"
             ],
