+++ run_test.ps1
@@ -2,4 +2,4 @@
 param (
   [switch] $CollectCoverage = $false
 )
-Get-ChildItem test/*Test.csproj -Recurse | ForEach-Object { dotnet test $_.FullName /p:CollectCoverage=$CollectCoverage }
+Get-ChildItem test/*Tests.csproj -Recurse | ForEach-Object { dotnet test $_.FullName /p:CollectCoverage=$CollectCoverage }
