+++ src/Directory.Build.props
@@ -1,8 +1,8 @@
 <Project>
   <PropertyGroup>
-    <Version Condition=" '$(Version)'=='' ">1.0.1</Version>
-    <PackageReleaseNotes>1.0.1 : 
-      &#xD;&#xA;Add ASP.NET Core Options pattern configuration package.
+    <Version Condition=" '$(Version)'=='' ">1.0.2</Version>
+    <PackageReleaseNotes>1.0.2 : 
+      &#xD;&#xA;Update Microsoft.Extensions.Options.ConfigurationExtensions package to .NET 5
       &#xD;&#xA;
     </PackageReleaseNotes>
   </PropertyGroup>
